Merge BigDecimal

+++ calculator_publihser/src/main/java/com/calculator_publisher/controller/domain/Resultado.java
@@ -1,22 +1,24 @@
 package com.calculator_publisher.controller.domain;
 
+import java.math.BigDecimal;
+
 /**
  * @Author Hanica Munguambe
- * @Author 13/11/2021 17:20
+ * @Author 14/11/2021 23:20
  */
 public class Resultado {
 
-    private float resultado;
+    private BigDecimal resultado;
 
     public Resultado() {}
 
-    public Resultado(float resultado){ this.resultado=resultado;}
+    public Resultado(BigDecimal resultado){ this.resultado=resultado;}
 
-    public float getResultado() {
+    public BigDecimal getResultado() {
         return resultado;
     }
 
-    public void setResultado(float resultado) {
+    public void setResultado(BigDecimal resultado) {
         this.resultado = resultado;
     }
 }
